Move customRenderHook in another file
