+++ packages/react-day-picker/src/test/customRender.tsx
@@ -1,10 +1,5 @@
 import * as React from 'react';
 import { render, RenderResult } from '@testing-library/react';
-import {
-  renderHook,
-  Renderer,
-  RenderHookResult
-} from '@testing-library/react-hooks';
 
 import { DayPickerProps } from 'types';
 
@@ -15,14 +10,4 @@ export const customRender = (
   contextValue: DayPickerProps = {}
 ): RenderResult => {
   return render(<ContextProvider {...contextValue}>{element}</ContextProvider>);
-};
-
-export const customRenderHook = <TProps, TResult>(
-  callback: (props: TProps) => TResult,
-  contextValue: DayPickerProps = {}
-): RenderHookResult<TProps, TResult, Renderer<TProps>> => {
-  const wrapper = ({ children }: { children?: React.ReactNode }) => (
-    <ContextProvider {...contextValue}>{children}</ContextProvider>
-  );
-  return renderHook<TProps, TResult>(callback, { wrapper });
 };
