+++ packages/react-day-picker/src/test/index.ts
@@ -1,2 +1,3 @@
 export * from './customRender';
+export * from './customRenderHook';
 export * from './PageObjects';
